test(MenuPriceTest.java) : 메뉴의 가격은 0원 이상이어야 한다.

+++ src/test/java/kitchenpos/menus/tobe/domain/menu/MenuPriceTest.java
@@ -3,6 +3,8 @@ package kitchenpos.menus.tobe.domain.menu;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class MenuPriceTest {
@@ -10,7 +12,7 @@ class MenuPriceTest {
     @DisplayName("메뉴 가격은 0원 이하일 수 없다.")
     @Test
     void createMenuPrice() {
-        assertThatThrownBy(() -> new MenuPrice(-1))
+        assertThatThrownBy(() -> new MenuPrice(BigDecimal.valueOf(-1)))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("메뉴 가격은 0원 이하일 수 없습니다.");
     }
